Refacto CRUD cart rule tests
