+++ tests/UI/campaigns/functional/BO/03_catalog/07_discounts/01_cartRules/01_CRUDCartRule/02_condition/10_countryRestriction.ts
@@ -11,6 +11,8 @@ import loginCommon from '@commonTests/BO/loginBO';
 import cartRulesPage from '@pages/BO/catalog/discounts';
 import addCartRulePage from '@pages/BO/catalog/discounts/add';
 import dashboardPage from '@pages/BO/dashboard';
+import zonesPage from '@pages/BO/international/locations';
+import countriesPage from '@pages/BO/international/locations/countries';
 // Import FO pages
 import cartPage from '@pages/FO/cart';
 import checkoutPage from '@pages/FO/checkout';
@@ -18,7 +20,7 @@ import foHomePage from '@pages/FO/home';
 import foProductPage from '@pages/FO/product';
 
 // Import data
-import Carriers from '@data/demo/carriers';
+import Countries from '@data/demo/countries';
 import Customers from '@data/demo/customers';
 import Products from '@data/demo/products';
 import CartRuleData from '@data/faker/cartRule';
@@ -26,22 +28,19 @@ import CartRuleData from '@data/faker/cartRule';
 import {expect} from 'chai';
 import type {BrowserContext, Page} from 'playwright';
 
-const baseContext: string = 'functional_BO_catalog_discounts_cartRules_CRUDCartRule_carrierRestriction';
+const baseContext: string = 'functional_BO_catalog_discounts_cartRules_CRUDCartRule_condition_countryRestriction';
 
-describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () => {
+describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () => {
   let browserContext: BrowserContext;
   let page: Page;
 
-  const cartRuleCode: CartRuleData = new CartRuleData({
+  const cartRule: CartRuleData = new CartRuleData({
     name: 'addCartRuleName',
     code: '4QABV6L3',
-    carrierRestriction: true,
+    countrySelection: true,
+    countryIDToRemove: Countries.france.id,
     discountType: 'Amount',
-    discountAmount: {
-      value: 10,
-      currency: 'EUR',
-      tax: 'Tax included',
-    },
+    discountPercent: 100,
   });
 
   // before and after functions
@@ -54,18 +53,75 @@ describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () =
     await helper.closeBrowserContext(browserContext);
   });
 
-  describe('Create a Cart Rules', async () => {
-    it('should login in BO', async function () {
-      await loginCommon.loginBO(this, page);
+  it('should login in BO', async function () {
+    await loginCommon.loginBO(this, page);
+  });
+
+  describe(`Enable the country '${Countries.unitedStates.name}'`, async () => {
+    it('should go to \'International > Locations\' page', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'goToLocationsPage', baseContext);
+
+      await dashboardPage.goToSubMenu(
+        page,
+        dashboardPage.internationalParentLink,
+        dashboardPage.locationsLink,
+      );
+      await zonesPage.closeSfToolBar(page);
+
+      const pageTitle = await zonesPage.getPageTitle(page);
+      await expect(pageTitle).to.contains(zonesPage.pageTitle);
     });
 
+    it('should go to \'Countries\' page', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'goToCountriesPage', baseContext);
+
+      await zonesPage.goToSubTabCountries(page);
+
+      const pageTitle = await countriesPage.getPageTitle(page);
+      await expect(pageTitle).to.contains(countriesPage.pageTitle);
+    });
+
+    it('should reset all filters and get number of countries in BO', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'resetFilterFirst', baseContext);
+
+      const numberOfCountries = await countriesPage.resetAndGetNumberOfLines(page);
+      await expect(numberOfCountries).to.be.above(0);
+    });
+
+    it(`should search for the country '${Countries.unitedStates.name}'`, async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'filterByNameToEnable', baseContext);
+
+      await countriesPage.filterTable(page, 'input', 'b!name', Countries.unitedStates.name);
+
+      const numberOfCountriesAfterFilter = await countriesPage.getNumberOfElementInGrid(page);
+      await expect(numberOfCountriesAfterFilter).to.be.equal(1);
+
+      const textColumn = await countriesPage.getTextColumnFromTable(page, 1, 'b!name');
+      await expect(textColumn).to.equal(Countries.unitedStates.name);
+    });
+
+    it('should enable the country', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'verifyTheCountryStatus', baseContext);
+
+      await countriesPage.setCountryStatus(page, 1, true);
+
+      const currentStatus = await countriesPage.getCountryStatus(page, 1);
+      await expect(currentStatus).to.be.true;
+    });
+
+    it('should reset filter', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'resetFilterCountry', baseContext);
+
+      await countriesPage.resetFilter(page);
+
+      const numberOfCountriesAfterReset = await countriesPage.resetAndGetNumberOfLines(page);
+      await expect(numberOfCountriesAfterReset).to.be.at.least(1);
+    });
+  });
+
+  describe('Create new cart rule with country restriction', async () => {
     it('should go to \'Catalog > Discounts\' page', async function () {
-      await testContext.addContextItem(
-        this,
-        'testIdentifier',
-        'goToDiscountsPage',
-        baseContext,
-      );
+      await testContext.addContextItem(this, 'testIdentifier', 'goToDiscountsPage', baseContext);
 
       await dashboardPage.goToSubMenu(
         page,
@@ -78,12 +134,7 @@ describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () =
     });
 
     it('should go to new cart rule page', async function () {
-      await testContext.addContextItem(
-        this,
-        'testIdentifier',
-        'goToNewCartRulePage',
-        baseContext,
-      );
+      await testContext.addContextItem(this, 'testIdentifier', 'goToNewCartRulePage', baseContext);
 
       await cartRulesPage.goToAddNewCartRulesPage(page);
 
@@ -92,24 +143,16 @@ describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () =
     });
 
     it('should create cart rule', async function () {
-      await testContext.addContextItem(
-        this,
-        'testIdentifier',
-        'createCartRule',
-        baseContext,
-      );
+      await testContext.addContextItem(this, 'testIdentifier', 'createCartRule', baseContext);
 
-      const validationMessage = await addCartRulePage.createEditCartRules(page, cartRuleCode);
+      const validationMessage = await addCartRulePage.createEditCartRules(page, cartRule);
       await expect(validationMessage).to.contains(addCartRulePage.successfulCreationMessage);
     });
+  });
 
+  describe('Check the created discount in FO', async () => {
     it('should view my shop', async function () {
-      await testContext.addContextItem(
-        this,
-        'testIdentifier',
-        'viewMyShop1',
-        baseContext,
-      );
+      await testContext.addContextItem(this, 'testIdentifier', 'viewMyShop1', baseContext);
 
       page = await addCartRulePage.viewMyShop(page);
       await foHomePage.changeLanguage(page, 'en');
@@ -117,9 +160,7 @@ describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () =
       const isHomePage = await foHomePage.isHomePage(page);
       await expect(isHomePage, 'Fail to open FO home page').to.be.true;
     });
-  });
 
-  describe('Use Cart Rule', async () => {
     it('should go to the first product page', async function () {
       await testContext.addContextItem(
         this,
@@ -148,22 +189,21 @@ describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () =
       await expect(notificationsNumber).to.be.equal(1);
     });
 
-    it('should add the promo code and get "You must choose a carrier before applying this voucher" error message',
-      async function () {
-        await testContext.addContextItem(
-          this,
-          'testIdentifier',
-          'addPromoCodeAndGetErrorMessage',
-          baseContext,
-        );
+    it('should set the promo code and verify the error message', async function () {
+      await testContext.addContextItem(
+        this,
+        'testIdentifier',
+        'addPromoCode',
+        baseContext,
+      );
 
-        await cartPage.addPromoCode(page, cartRuleCode.code);
+      await cartPage.addPromoCode(page, cartRule.code);
 
-        const alertMessage = await cartPage.getCartRuleErrorMessage(page);
-        await expect(alertMessage).to.equal(cartPage.cartRuleChooseCarrierAlertMessageText);
-      });
+      const chooseDeliveryAddressNotification = await cartPage.getAlertWarningForPromoCode(page);
+      await expect(chooseDeliveryAddressNotification).to.equal(cartPage.alertChooseDeliveryAddressWarningText);
+    });
 
-    it('should proceed to checkouts', async function () {
+    it('should proceed to checkout', async function () {
       await testContext.addContextItem(
         this,
         'testIdentifier',
@@ -171,19 +211,15 @@ describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () =
         baseContext,
       );
 
+      // Proceed to checkout the shopping cart
       await cartPage.clickOnProceedToCheckout(page);
 
       const isCheckout = await checkoutPage.isCheckoutPage(page);
       await expect(isCheckout).to.be.true;
     });
 
-    it('should be authentificated', async function () {
-      await testContext.addContextItem(
-        this,
-        'testIdentifier',
-        'signInFO',
-        baseContext,
-      );
+    it('should checkout by signIn', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'signInFO', baseContext);
 
       await checkoutPage.clickOnSignIn(page);
 
@@ -191,31 +227,18 @@ describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () =
       await expect(isCustomerConnected, 'Customer is not connected').to.be.true;
     });
 
-    it('should confirm adress after signIn', async function () {
+    it('should confirm address after signIn', async function () {
       await testContext.addContextItem(
         this,
         'testIdentifier',
-        'confirmAddressStep',
+        'chooseAndConfirmAddressStep',
         baseContext,
       );
+
+      await checkoutPage.chooseDeliveryAddress(page, 2);
 
       const isDeliveryStep = await checkoutPage.goToDeliveryStep(page);
       await expect(isDeliveryStep, 'Delivery Step block is not displayed').to.be.true;
-    });
-
-    it('should set the promo code and choose the wrong shipping method', async function () {
-      await testContext.addContextItem(
-        this,
-        'testIdentifier',
-        'addPromoCodeAndChooseWrongShippingMethod',
-        baseContext,
-      );
-
-      await checkoutPage.chooseShippingMethodAndAddComment(page, 1);
-      await checkoutPage.addPromoCode(page, cartRuleCode.code);
-
-      const errorShippingMessage = await checkoutPage.getCartRuleErrorMessage(page);
-      await expect(errorShippingMessage).to.equal(cartPage.cartRuleCannotUseVoucherAlertMessageText);
     });
 
     it('should set the promo code for second time and check total after discount', async function () {
@@ -226,24 +249,14 @@ describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () =
         baseContext,
       );
 
-      await checkoutPage.goToShippingStep(page);
-      await checkoutPage.chooseShippingMethodAndAddComment(page, 2);
-      await checkoutPage.addPromoCode(page, cartRuleCode.code);
+      await checkoutPage.addPromoCode(page, cartRule.code);
 
-      let discountedPrice = cartRuleCode.discountAmount.value;
-
-      if (discountedPrice >= Products.demo_1.finalPrice) {
-        discountedPrice = Products.demo_1.finalPrice;
-      }
-
-      const priceATI = await checkoutPage.getATIPrice(page);
-      await expect(priceATI).to.equal(parseFloat(
-        (Products.demo_1.finalPrice - discountedPrice + Carriers.myCarrier.priceTTC)
-          .toFixed(2),
-      ),
-      );
+      const cartRuleName = await cartPage.getCartRuleName(page, 1);
+      await expect(cartRuleName).to.equal(cartRule.name);
     });
+  });
 
+  describe('Delete the shopping cart', async () => {
     it('should remove the discount', async function () {
       await testContext.addContextItem(
         this,
@@ -256,7 +269,7 @@ describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () =
       await expect(isDeleteIconNotVisible, 'The discount is not removed').to.be.true;
     });
 
-    it('should click on the logo of the shop', async function () {
+    it('should go to Home page', async function () {
       await testContext.addContextItem(
         this,
         'testIdentifier',
@@ -297,5 +310,5 @@ describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () =
   });
 
   // post condition : delete cart rule
-  deleteCartRuleTest(cartRuleCode.name, `${baseContext}_postTest_1`);
+  deleteCartRuleTest(cartRule.name, `${baseContext}_postTest_1`);
 });
