+++ tests/UI/campaigns/functional/BO/03_catalog/07_discounts/01_cartRules/01_CRUDCartRule/02_condition/11_carrierRestriction.ts
@@ -11,8 +11,6 @@ import loginCommon from '@commonTests/BO/loginBO';
 import cartRulesPage from '@pages/BO/catalog/discounts';
 import addCartRulePage from '@pages/BO/catalog/discounts/add';
 import dashboardPage from '@pages/BO/dashboard';
-import zonesPage from '@pages/BO/international/locations';
-import countriesPage from '@pages/BO/international/locations/countries';
 // Import FO pages
 import cartPage from '@pages/FO/cart';
 import checkoutPage from '@pages/FO/checkout';
@@ -20,7 +18,7 @@ import foHomePage from '@pages/FO/home';
 import foProductPage from '@pages/FO/product';
 
 // Import data
-import Countries from '@data/demo/countries';
+import Carriers from '@data/demo/carriers';
 import Customers from '@data/demo/customers';
 import Products from '@data/demo/products';
 import CartRuleData from '@data/faker/cartRule';
@@ -28,19 +26,22 @@ import CartRuleData from '@data/faker/cartRule';
 import {expect} from 'chai';
 import type {BrowserContext, Page} from 'playwright';
 
-const baseContext: string = 'functional_BO_catalog_discounts_cartRules_CRUDCartRule_countryRestriction';
+const baseContext: string = 'functional_BO_catalog_discounts_cartRules_CRUDCartRule_condition_carrierRestriction';
 
-describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () => {
+describe('BO - Catalog - Cart rules : Case 11 - Carrier Restriction', async () => {
   let browserContext: BrowserContext;
   let page: Page;
 
-  const cartRule: CartRuleData = new CartRuleData({
+  const cartRuleCode: CartRuleData = new CartRuleData({
     name: 'addCartRuleName',
     code: '4QABV6L3',
-    countrySelection: true,
-    countryIDToRemove: Countries.france.id,
+    carrierRestriction: true,
     discountType: 'Amount',
-    discountPercent: 100,
+    discountAmount: {
+      value: 10,
+      currency: 'EUR',
+      tax: 'Tax included',
+    },
   });
 
   // before and after functions
@@ -53,75 +54,18 @@ describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () =
     await helper.closeBrowserContext(browserContext);
   });
 
-  it('should login in BO', async function () {
-    await loginCommon.loginBO(this, page);
-  });
-
-  describe(`Enable the country '${Countries.unitedStates.name}'`, async () => {
-    it('should go to \'International > Locations\' page', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'goToLocationsPage', baseContext);
-
-      await dashboardPage.goToSubMenu(
-        page,
-        dashboardPage.internationalParentLink,
-        dashboardPage.locationsLink,
-      );
-      await zonesPage.closeSfToolBar(page);
-
-      const pageTitle = await zonesPage.getPageTitle(page);
-      await expect(pageTitle).to.contains(zonesPage.pageTitle);
+  describe('Create a Cart Rules', async () => {
+    it('should login in BO', async function () {
+      await loginCommon.loginBO(this, page);
     });
 
-    it('should go to \'Countries\' page', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'goToCountriesPage', baseContext);
-
-      await zonesPage.goToSubTabCountries(page);
-
-      const pageTitle = await countriesPage.getPageTitle(page);
-      await expect(pageTitle).to.contains(countriesPage.pageTitle);
-    });
-
-    it('should reset all filters and get number of countries in BO', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'resetFilterFirst', baseContext);
-
-      const numberOfCountries = await countriesPage.resetAndGetNumberOfLines(page);
-      await expect(numberOfCountries).to.be.above(0);
-    });
-
-    it(`should search for the country '${Countries.unitedStates.name}'`, async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'filterByNameToEnable', baseContext);
-
-      await countriesPage.filterTable(page, 'input', 'b!name', Countries.unitedStates.name);
-
-      const numberOfCountriesAfterFilter = await countriesPage.getNumberOfElementInGrid(page);
-      await expect(numberOfCountriesAfterFilter).to.be.equal(1);
-
-      const textColumn = await countriesPage.getTextColumnFromTable(page, 1, 'b!name');
-      await expect(textColumn).to.equal(Countries.unitedStates.name);
-    });
-
-    it('should enable the country', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'verifyTheCountryStatus', baseContext);
-
-      await countriesPage.setCountryStatus(page, 1, true);
-
-      const currentStatus = await countriesPage.getCountryStatus(page, 1);
-      await expect(currentStatus).to.be.true;
-    });
-
-    it('should reset filter', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'resetFilterCountry', baseContext);
-
-      await countriesPage.resetFilter(page);
-
-      const numberOfCountriesAfterReset = await countriesPage.resetAndGetNumberOfLines(page);
-      await expect(numberOfCountriesAfterReset).to.be.at.least(1);
-    });
-  });
-
-  describe('Create new cart rule with country restriction', async () => {
     it('should go to \'Catalog > Discounts\' page', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'goToDiscountsPage', baseContext);
+      await testContext.addContextItem(
+        this,
+        'testIdentifier',
+        'goToDiscountsPage',
+        baseContext,
+      );
 
       await dashboardPage.goToSubMenu(
         page,
@@ -134,7 +78,12 @@ describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () =
     });
 
     it('should go to new cart rule page', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'goToNewCartRulePage', baseContext);
+      await testContext.addContextItem(
+        this,
+        'testIdentifier',
+        'goToNewCartRulePage',
+        baseContext,
+      );
 
       await cartRulesPage.goToAddNewCartRulesPage(page);
 
@@ -143,16 +92,24 @@ describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () =
     });
 
     it('should create cart rule', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'createCartRule', baseContext);
+      await testContext.addContextItem(
+        this,
+        'testIdentifier',
+        'createCartRule',
+        baseContext,
+      );
 
-      const validationMessage = await addCartRulePage.createEditCartRules(page, cartRule);
+      const validationMessage = await addCartRulePage.createEditCartRules(page, cartRuleCode);
       await expect(validationMessage).to.contains(addCartRulePage.successfulCreationMessage);
     });
-  });
 
-  describe('Check the created discount in FO', async () => {
     it('should view my shop', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'viewMyShop1', baseContext);
+      await testContext.addContextItem(
+        this,
+        'testIdentifier',
+        'viewMyShop1',
+        baseContext,
+      );
 
       page = await addCartRulePage.viewMyShop(page);
       await foHomePage.changeLanguage(page, 'en');
@@ -160,7 +117,9 @@ describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () =
       const isHomePage = await foHomePage.isHomePage(page);
       await expect(isHomePage, 'Fail to open FO home page').to.be.true;
     });
+  });
 
+  describe('Use Cart Rule', async () => {
     it('should go to the first product page', async function () {
       await testContext.addContextItem(
         this,
@@ -189,21 +148,22 @@ describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () =
       await expect(notificationsNumber).to.be.equal(1);
     });
 
-    it('should set the promo code and verify the error message', async function () {
-      await testContext.addContextItem(
-        this,
-        'testIdentifier',
-        'addPromoCode',
-        baseContext,
-      );
+    it('should add the promo code and get "You must choose a carrier before applying this voucher" error message',
+      async function () {
+        await testContext.addContextItem(
+          this,
+          'testIdentifier',
+          'addPromoCodeAndGetErrorMessage',
+          baseContext,
+        );
 
-      await cartPage.addPromoCode(page, cartRule.code);
+        await cartPage.addPromoCode(page, cartRuleCode.code);
 
-      const chooseDeliveryAddressNotification = await cartPage.getAlertWarningForPromoCode(page);
-      await expect(chooseDeliveryAddressNotification).to.equal(cartPage.alertChooseDeliveryAddressWarningText);
-    });
+        const alertMessage = await cartPage.getCartRuleErrorMessage(page);
+        await expect(alertMessage).to.equal(cartPage.cartRuleChooseCarrierAlertMessageText);
+      });
 
-    it('should proceed to checkout', async function () {
+    it('should proceed to checkouts', async function () {
       await testContext.addContextItem(
         this,
         'testIdentifier',
@@ -211,15 +171,19 @@ describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () =
         baseContext,
       );
 
-      // Proceed to checkout the shopping cart
       await cartPage.clickOnProceedToCheckout(page);
 
       const isCheckout = await checkoutPage.isCheckoutPage(page);
       await expect(isCheckout).to.be.true;
     });
 
-    it('should checkout by signIn', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'signInFO', baseContext);
+    it('should be authentificated', async function () {
+      await testContext.addContextItem(
+        this,
+        'testIdentifier',
+        'signInFO',
+        baseContext,
+      );
 
       await checkoutPage.clickOnSignIn(page);
 
@@ -227,18 +191,31 @@ describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () =
       await expect(isCustomerConnected, 'Customer is not connected').to.be.true;
     });
 
-    it('should confirm address after signIn', async function () {
+    it('should confirm adress after signIn', async function () {
       await testContext.addContextItem(
         this,
         'testIdentifier',
-        'chooseAndConfirmAddressStep',
+        'confirmAddressStep',
         baseContext,
       );
 
-      await checkoutPage.chooseDeliveryAddress(page, 2);
-
       const isDeliveryStep = await checkoutPage.goToDeliveryStep(page);
       await expect(isDeliveryStep, 'Delivery Step block is not displayed').to.be.true;
+    });
+
+    it('should set the promo code and choose the wrong shipping method', async function () {
+      await testContext.addContextItem(
+        this,
+        'testIdentifier',
+        'addPromoCodeAndChooseWrongShippingMethod',
+        baseContext,
+      );
+
+      await checkoutPage.chooseShippingMethodAndAddComment(page, 1);
+      await checkoutPage.addPromoCode(page, cartRuleCode.code);
+
+      const errorShippingMessage = await checkoutPage.getCartRuleErrorMessage(page);
+      await expect(errorShippingMessage).to.equal(cartPage.cartRuleCannotUseVoucherAlertMessageText);
     });
 
     it('should set the promo code for second time and check total after discount', async function () {
@@ -249,14 +226,24 @@ describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () =
         baseContext,
       );
 
-      await checkoutPage.addPromoCode(page, cartRule.code);
+      await checkoutPage.goToShippingStep(page);
+      await checkoutPage.chooseShippingMethodAndAddComment(page, 2);
+      await checkoutPage.addPromoCode(page, cartRuleCode.code);
 
-      const cartRuleName = await cartPage.getCartRuleName(page, 1);
-      await expect(cartRuleName).to.equal(cartRule.name);
+      let discountedPrice = cartRuleCode.discountAmount.value;
+
+      if (discountedPrice >= Products.demo_1.finalPrice) {
+        discountedPrice = Products.demo_1.finalPrice;
+      }
+
+      const priceATI = await checkoutPage.getATIPrice(page);
+      await expect(priceATI).to.equal(parseFloat(
+        (Products.demo_1.finalPrice - discountedPrice + Carriers.myCarrier.priceTTC)
+          .toFixed(2),
+      ),
+      );
     });
-  });
 
-  describe('Delete the shopping cart', async () => {
     it('should remove the discount', async function () {
       await testContext.addContextItem(
         this,
@@ -269,7 +256,7 @@ describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () =
       await expect(isDeleteIconNotVisible, 'The discount is not removed').to.be.true;
     });
 
-    it('should go to Home page', async function () {
+    it('should click on the logo of the shop', async function () {
       await testContext.addContextItem(
         this,
         'testIdentifier',
@@ -310,5 +297,5 @@ describe('BO - Catalog - Cart rules : Case 10 - Country Restriction', async () =
   });
 
   // post condition : delete cart rule
-  deleteCartRuleTest(cartRule.name, `${baseContext}_postTest_1`);
+  deleteCartRuleTest(cartRuleCode.name, `${baseContext}_postTest_1`);
 });
